Commit `Cargo.lock` and set default Rust toolchain to 1.70.0
Per the updated guidance from the Cargo team, library crates can check
in a lockfile if they want to test against their MSRV. And in any case,
atrium-cli is a binary crate, so it needs a lockfile to be checked in.
The lockfile pins dependency versions that work with 1.70 (in particular
`clap < 4.5`).

Committing a `rust-toolchain.toml` file will result in CI running
against MSRV by default, and developers using the MSRV toolchain by
default (avoiding accidentally depending on features in later compiler
versions).